Should no longer send you back 2 pages if not coming from level page
The reason I used sessionStorage is becuase there's no way to access a list of previously visited URLs (for security reasons) so basically, when a page is about to unload it saves the current URL. When a level download page is loaded, it checks if the saved URL is the level page one. After gdbrowser is closed saved URL goes away.